requirements-llava.txt
accelerate==0.21.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.18
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
einops==0.6.1
einops-exts==0.0.4
executing==2.2.1
fastapi==0.124.0
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
gradio==3.35.2
gradio-client==0.2.9
h11==0.14.0
hf-xet==1.2.0
httpcore
httpx
huggingface-hub==0.36.0
idna==3.11
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.3.1
latex2mathml==3.78.1
linkify-it-py==2.0.3
markdown-it-py==2.2.0
markdown2==2.5.4
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
narwhals==2.13.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==1.12.0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
peft==0.4.0
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.23
pydantic==1.10.24
pydub==0.25.1
pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.2.2
scipy==1.16.3
semantic-version==2.10.0
send2trash==1.8.3
sentencepiece==0.1.99
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.50.0
svgwrite==1.4.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==0.9.12
tinycss2==1.4.0
tokenizers==0.15.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.36.2
triton==3.5.1
typing-extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.6.1
uvicorn==0.38.0
wavedrom==2.0.3.post3
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
yarl==1.22.0
